+++ mycodo/mycodo_flask/forms/forms_settings.py
@@ -139,6 +139,7 @@ class SettingsGeneral(FlaskForm):
     landing_page = StringField(lazy_gettext('Landing Page'))
     language = StringField(lazy_gettext('Language'))
     rpyc_timeout = StringField(lazy_gettext('Pyro Timeout'))
+    daemon_debug_mode = BooleanField(lazy_gettext('Enable Daemon Debug Logging'))
     force_https = BooleanField(lazy_gettext('Force HTTPS'))
     hide_success = BooleanField(lazy_gettext('Hide success messages'))
     hide_info = BooleanField(lazy_gettext('Hide info messages'))